force re-shuffle if some items return to their original spot
since the point of the activity is to give learners practice putting lines in the correct order, it makes it artificially easier if some of the lines end up back in the same place they started (meaning part of the "correct" answer is already supplied to the student).

So I added a very hacky recursive loop to force a re-shuffle if any of the lines end up back in their original position after the initial shuffle.